Add ViZDoomGym to list of tools (#450)
Thanks for the PR! Gym-wrapper for ViZDoom is always welcome :).

One suggestion for your repo though: Add support for getting `GameVariables` as defined in the config file. For some scenarios these are necessary for the agent to be able to learn the scenario. You can not solve `health_gathering_supreme.cfg` without `HEALTH` variable (or some "time survived" variable), for example.